* Replace existed random generators by Mersenne Twister random number generator. Cleanup code.
git-svn-id: http://mangos.svn.sourceforge.net/svnroot/mangos@5998 26674162-90ff-0310-a85b-b64107c8dee9

+++ trunk/dep/src/sockets/Utility.cpp
@@ -952,7 +952,7 @@ std::string Utility::ToString(double d)
 
 unsigned long Utility::Rnd()
 {
-static	Utility::Rng generator( time(NULL) );
+static	Utility::Rng generator( (unsigned long)time(NULL) );
 	return generator.Get();
 }
